@@ -6,5 +6,5 @@ import "math"
 
 // GetNumFDs returns the File Descriptors for non unix systems as MaxInt.
 func GetNumFDs() uint64 {
-	return math.MaxUint64
+	return uint64(math.MaxUint64)
 }
