@@ -118,14 +118,7 @@ func TestMain(m *testing.M) {
 	}
 
 	for _, f := range customLogLvlFacilities {
-		if _, ok := LoggingFacilities[f]; ok {
-			SetFacilityLogLevel(f, logLevel)
-			continue
-		}
-		if _, ok := LoggingFacilitiesExtra[f]; ok {
-			SetFacilityLogLevel(f, logLevel)
-			continue
-		}
+		SetFacilityLogLevel(f, logLevel)
 	}
 
 	diskInfCfg := &disk.Config{}
@@ -302,7 +295,7 @@ func makeConsensus(t *testing.T, store ds.Datastore, h host.Host, psub *pubsub.P
 }
 
 func createCluster(t *testing.T, host host.Host, dht *dual.DHT, clusterCfg *Config, store ds.Datastore, consensus Consensus, apis []API, ipfs IPFSConnector, tracker PinTracker, mon PeerMonitor, alloc PinAllocator, inf Informer, tracer Tracer) *Cluster {
-	cl, err := NewCluster(context.Background(), host, dht, clusterCfg, store, consensus, apis, ipfs, tracker, mon, alloc, []Informer{inf}, tracer)
+	cl, err := NewCluster(context.Background(), host, nil, dht, clusterCfg, store, consensus, apis, ipfs, tracker, mon, alloc, []Informer{inf}, tracer)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -355,7 +348,11 @@ func createHost(t *testing.T, priv crypto.PrivKey, clusterSecret []byte, listen 
 
 	// Pubsub needs to be created BEFORE connecting the peers,
 	// otherwise they are not picked up.
-	psub, err := newPubSub(ctx, h)
+	// Note: this is possibly a pubsub bug that was fixed.
+	cfg := &Config{}
+	cfg.Default()
+	cfg.LoadJSON(testingClusterCfg)
+	psub, err := newPubSub(ctx, cfg, h)
 	if err != nil {
 		t.Fatal(err)
 	}
