@@ -10,6 +10,7 @@ import (
 	"io"
 	"net/http"
 	"net/url"
+	"os"
 	"strconv"
 	"strings"
 	"sync"
@@ -415,6 +416,9 @@ func (ipfs *Connector) Pin(ctx context.Context, pin api.Pin) error {
 		return nil
 	}
 
+	// Call at the beginning of pinning to update pinqueue
+	ipfs.updateInformerMetric(ctx)
+	// Call at the end of pinning to update freespace
 	defer ipfs.updateInformerMetric(ctx)
 
 	ctx, cancelRequest := context.WithCancel(ctx)
@@ -568,7 +572,9 @@ func (ipfs *Connector) Unpin(ctx context.Context, hash api.Cid) error {
 		return errors.New("ipfs unpinning is disallowed by configuration on this peer")
 	}
 
-	defer ipfs.updateInformerMetric(ctx)
+	// Unpinning doesn't free space and doesn't matter for pinqueue so not
+	// really necessary to publish metrics.
+	//defer ipfs.updateInformerMetric(ctx)
 
 	path := fmt.Sprintf("pin/rm?arg=%s", hash)
 
@@ -829,6 +835,9 @@ func (ipfs *Connector) RepoGC(ctx context.Context) (api.RepoGC, error) {
 
 	defer body.Close()
 
+	// Freespace metric might have gone down, so update it at the end.
+	defer ipfs.updateInformerMetric(ctx)
+
 	dec := json.NewDecoder(body)
 	repoGC := api.RepoGC{
 		Keys: []api.IPFSRepoGC{},
@@ -949,6 +958,16 @@ func (cd *chanDirectory) Size() (int64, error) {
 
 func (cd *chanDirectory) Entries() files.DirIterator {
 	return cd.iterator
+}
+
+// Mode: return mode for directory, but unused.
+func (cd *chanDirectory) Mode() os.FileMode {
+	return os.ModeDir
+}
+
+// ModeTime: not implemented
+func (cd *chanDirectory) ModTime() (mtime time.Time) {
+	return time.UnixMilli(0)
 }
 
 // chanIterator implements the files.DirIterator interface.
@@ -1096,6 +1115,9 @@ func (ipfs *Connector) BlockStream(ctx context.Context, blocks <-chan api.NodeWi
 	defer span.End()
 
 	logger.Debug("streaming blocks to IPFS")
+
+	// Update at the end of block-streaming to have an updated freespace
+	// metric.
 	defer ipfs.updateInformerMetric(ctx)
 
 	it := &chanIterator{
@@ -1202,7 +1224,7 @@ func (ipfs *Connector) shouldUpdateMetric() bool {
 		return false
 	}
 	curCount := atomic.AddUint64(&ipfs.updateMetricCount, 1)
-	if curCount%uint64(ipfs.config.InformerTriggerInterval) == 0 {
+	if curCount >= uint64(ipfs.config.InformerTriggerInterval) {
 		atomic.StoreUint64(&ipfs.updateMetricCount, 0)
 		return true
 	}
