@@ -7,7 +7,7 @@ import (
 )
 
 // Version is the current cluster version.
-var Version = semver.MustParse("1.1.0")
+var Version = semver.MustParse("1.1.2")
 
 // RPCProtocol is protocol handler used to send libp2p-rpc messages between
 // cluster peers.  All peers in the cluster need to speak the same protocol
